Have awk indexer output filetags.
From what I'd tested, the awk indexer doesn't seem to parse the
 #+FILETAGS keyword properly if it is under a file-level property drawer
but it works as expected when the keyword is at the first line of the
file. I guess the problem lies in the ordering of patterns in the awk
code and I have to figure it out. For now, I have the indexer
outputs the filetags property when the file processing ended so that
there is no tags left behind and this filetags property will be
concatenated with the inherited-tags property. Tags duplication may occur
but it will be dealt with.